+++ src/reportlab/__init__.py
@@ -1,9 +1,9 @@
 #Copyright ReportLab Europe Ltd. 2000-2023
 #see license.txt for license details
 __doc__="""The Reportlab PDF generation library."""
-Version = "4.4.9"
+Version = "4.4.10"
 __version__=Version
-__date__='20260115'
+__date__='20260212'
 
 import sys, os